light refactoring based on github review.
