+++ src/modules/reporting/selectors/get-payout-numerators.js
